pyproject.toml
[tool.poetry]
name = "budget-notify"
version = "1.0.0"
description = ""
authors = ["Mohammed Rishad <mohammedrishad.vk@gmail.com>"]
readme = "README.md"
packages = [{include = "budget_notify"}]

[tool.poetry.dependencies]
python = "^3.9"
mysql-connector-python = "^8.0.32"
pytest = "^7.2.1"

[tool.black]
line-length = 88
include = '\.pyi?$'
exclude = '''
/(
    \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
  | migrations
)/
'''

[tool.isort]
# Make sure this matches `*.py` in .editorconfig
ensure_newline_before_comments = true
force_grid_wrap = 0
include_trailing_comma = true
line_length = 88
multi_line_output = 3
use_parentheses = true


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
